refactor: LineGenerator -> Line 역할 수정
