+++ src/test/java/ladder/domain/LineTest.java
@@ -1,4 +1,4 @@
-package ladder.domain.generator;
+package ladder.domain;
 
 import static ladder.domain.Direction.LEFT;
 import static ladder.domain.Direction.RIGHT;
@@ -7,22 +7,21 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import java.util.stream.Stream;
-import ladder.domain.Direction;
-import ladder.domain.Line;
+import ladder.domain.generator.DirectionGenerator;
+import ladder.domain.generator.TestDirectionGenerator;
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class LineGeneratorTest {
+class LineTest {
 
     private static Stream<Arguments> generateDirections() {
         return Stream.of(
                 Arguments.of(Lists.newArrayList(RIGHT, STAY, STAY), new Direction[]{RIGHT, LEFT, STAY, STAY}, 4),
                 Arguments.of(Lists.newArrayList(RIGHT, RIGHT), new Direction[]{RIGHT, LEFT, RIGHT, LEFT}, 4),
-                Arguments.of(Lists.newArrayList(RIGHT, STAY), new Direction[]{RIGHT, LEFT, STAY}, 3),
-                Arguments.of(Lists.newArrayList(RIGHT), new Direction[]{STAY}, 1)
+                Arguments.of(Lists.newArrayList(RIGHT, STAY), new Direction[]{RIGHT, LEFT, STAY}, 3)
         );
     }
 
@@ -31,10 +30,8 @@ class LineGeneratorTest {
     @DisplayName("발판이 정상적으로 만들어진다.")
     void validLine(final List<Direction> directions, final Direction[] expected, final int directionCount) {
         final DirectionGenerator directionGenerator = new TestDirectionGenerator(directions);
-        final LineGenerator lineGenerator = new LineGenerator();
-        final Line result = lineGenerator.generate(directionGenerator, directionCount);
+        final Line line = Line.of(directionGenerator, directionCount);
 
-        assertThat(result.getDirections()).containsExactly(expected);
+        assertThat(line.getDirections()).containsExactly(expected);
     }
 }
-
